+++ packages/core/src/connectivity/scp-cf/destination/destination-accessor-authentication-type.spec.ts
@@ -33,7 +33,6 @@ import { wrapJwtInHeader } from '../jwt';
 import { parseDestination } from './destination';
 import { getDestination } from './destination-accessor';
 import { destinationCache } from './destination-cache';
-import { getDestinationFromDestinationService } from './destination-from-service';
 import {
   alwaysProvider,
   alwaysSubscriber
@@ -377,59 +376,6 @@ describe('authentication types', () => {
       });
       expect(actual).toMatchObject(expected);
       expect(serviceTokenSpy).toHaveBeenCalled();
-    });
-
-    it('retrieves destination without specifying userJwt', async () => {
-      mockServiceBindings();
-      mockServiceToken();
-
-      mockInstanceDestinationsCall(nock, [], 200, providerServiceToken);
-      mockSubaccountDestinationsCall(
-        nock,
-        basicMultipleResponse,
-        200,
-        providerServiceToken
-      );
-
-      const actual = await getDestination(destinationName, {
-        cacheVerificationKeys: false
-      });
-      const expected = parseDestination(basicMultipleResponse[0]);
-      expect(actual).toMatchObject(expected);
-    });
-
-    // Test for ONEmds specific feature
-    it('is possible to get a non-principal propagation destination by only providing the subdomain (iss) instead of the whole jwt', async () => {
-      mockServiceBindings();
-      mockServiceToken();
-
-      mockInstanceDestinationsCall(nock, [], 200, subscriberServiceToken);
-      mockSubaccountDestinationsCall(
-        nock,
-        certificateMultipleResponse,
-        200,
-        subscriberServiceToken
-      );
-      mockInstanceDestinationsCall(nock, [], 200, providerServiceToken);
-      mockSubaccountDestinationsCall(nock, [], 200, providerServiceToken);
-
-      mockSingleDestinationCall(
-        nock,
-        certificateSingleResponse,
-        200,
-        'ERNIE-UND-CERT',
-        wrapJwtInHeader(subscriberServiceToken).headers
-      );
-
-      const expected = parseDestination(certificateSingleResponse);
-      const actual = await getDestinationFromDestinationService(
-        'ERNIE-UND-CERT',
-        {
-          iss: 'https://subscriber.example.com',
-          cacheVerificationKeys: false
-        }
-      );
-      expect(actual).toMatchObject(expected);
     });
   });
 });
